added build-mat support.
This adds a '/builder/summary' page, which, when printed, will print a build-mat style printout. The summary page is a nice overview of the fleet, though the size of the cards probably needs to be reworked (they're a little big).

Additionally, several massive bugs with the squad builder have been fixed, including the wonky upgrades binding.